cobalt/syscall: fix in/out value types

+++ kernel/cobalt/trace/cobalt-posix.h
@@ -23,13 +23,13 @@
 	__entry->tv_sec_##__name, __entry->tv_nsec_##__name
 
 DECLARE_EVENT_CLASS(syscall_entry,
-	TP_PROTO(struct xnthread *thread, int nr),
+	TP_PROTO(struct xnthread *thread, unsigned int nr),
 	TP_ARGS(thread, nr),
 
 	TP_STRUCT__entry(
 		__field(struct xnthread *, thread)
 		__string(name, thread ? thread->name : "(anon)")
-		__field(int, nr)
+		__field(unsigned int, nr)
 	),
 
 	TP_fast_assign(
@@ -38,17 +38,17 @@ DECLARE_EVENT_CLASS(syscall_entry,
 		__entry->nr = nr;
 	),
 
-	TP_printk("thread=%p(%s) syscall=%d",
+	TP_printk("thread=%p(%s) syscall=%u",
 		  __entry->thread, __get_str(name), __entry->nr)
 );
 
 DECLARE_EVENT_CLASS(syscall_exit,
-	TP_PROTO(struct xnthread *thread, int result),
+	TP_PROTO(struct xnthread *thread, long result),
 	TP_ARGS(thread, result),
 
 	TP_STRUCT__entry(
 		__field(struct xnthread *, thread)
-		__field(int, result)
+		__field(long, result)
 	),
 
 	TP_fast_assign(
@@ -56,7 +56,7 @@ DECLARE_EVENT_CLASS(syscall_exit,
 		__entry->result = result;
 	),
 
-	TP_printk("thread=%p result=%d",
+	TP_printk("thread=%p result=%ld",
 		  __entry->thread, __entry->result)
 );
 
@@ -150,22 +150,22 @@ DECLARE_EVENT_CLASS(cobalt_void,
 );
 
 DEFINE_EVENT(syscall_entry, cobalt_head_sysentry,
-	TP_PROTO(struct xnthread *thread, int nr),
+	TP_PROTO(struct xnthread *thread, unsigned int nr),
 	TP_ARGS(thread, nr)
 );
 
 DEFINE_EVENT(syscall_exit, cobalt_head_sysexit,
-	TP_PROTO(struct xnthread *thread, int result),
+	TP_PROTO(struct xnthread *thread, long result),
 	TP_ARGS(thread, result)
 );
 
 DEFINE_EVENT(syscall_entry, cobalt_root_sysentry,
-	TP_PROTO(struct xnthread *thread, int nr),
+	TP_PROTO(struct xnthread *thread, unsigned int nr),
 	TP_ARGS(thread, nr)
 );
 
 DEFINE_EVENT(syscall_exit, cobalt_root_sysexit,
-	TP_PROTO(struct xnthread *thread, int result),
+	TP_PROTO(struct xnthread *thread, long result),
 	TP_ARGS(thread, result)
 );
